[1106 lines] pull out StubKubernetesManager into test_helper internal pkg
